Implemented the Huber loss, mean absolute error loss, changed metric calculations and changed the import structure
