+++ src/DLL/DeepLearning/Losses/__init__.py
@@ -1,3 +1,5 @@
 from .MSE import mse
+from .MAE import mae
 from .BCE import bce
 from .CCE import cce
+from .Huber import Huber
