--- src/DLL/DeepLearning/__init__.py
@@ -1,1 +1,0 @@
-from .Model import Model
